+++ src/pages/User/Profiles/DeleteProfile.tsx
