+++ src/pages/User/Profiles/index.tsx
@@ -1,0 +1,3 @@
+export * from './EditProfile';
+export * from './CreateProfile';
+export * from './Profiles';
